Remove old messenger
