+++ resources/VegaResources_de_DE.properties
@@ -740,8 +740,6 @@ ReadConfiguration_4K = Konfiguration lesen...
 
 RecentlyUsedEmailAddresses_4L = Bisher verwendete E-Mail-Adressen
 
-RefreshStatus_4M = Status aktualisieren
-
 Refresh_4N = Aktualisieren
 
 ReleaseFormatted_4O = {0}.{1}.{2} {3}:{4}
@@ -983,8 +981,6 @@ UnsentMessages2_EX = Es gibt ungesendete Nachrichten. M\u00F6chten Sie den Messe
 UnsentMessages_EW = Ungesendete Nachrichten
 
 UpdateUserQuestion_7S = M\u00F6chten Sie den User [{0}] wirklich aktualisieren?
-
-UserActicationQuestion_7T = M\u00F6chten Sie Ihren Benutzer [{0}] auf dem Server\n{1}:{2}\naktivieren? W\u00E4hlen Sie im folgenden Dialog den Ablageort f\u00FCr die Authentifizierungsdatei.
 
 UserActivationSuccess_7U = Der Benutzer wurde erfolgreich aktiviert.
