+++ resources/VegaResources_en_US.properties
@@ -740,8 +740,6 @@ ReadConfiguration_4K = Read configuration...
 
 RecentlyUsedEmailAddresses_4L = Recently used e-mail addresses
 
-RefreshStatus_4M = Refresh status
-
 Refresh_4N = Refresh
 
 ReleaseFormatted_4O = {1}/{0}/{2} {3}:{4}
@@ -983,8 +981,6 @@ UnsentMessages2_EX = There are unsent messages. Do you really want to close the 
 UnsentMessages_EW = Unsent messages
 
 UpdateUserQuestion_7S = Do you really want to update user [{0}]?
-
-UserActicationQuestion_7T = Do you want to acticate your user [{0}] on server\n{1}:{2}?\nSelect the storage location for the authentication file in the following dialog.
 
 UserActivationSuccess_7U = The user was activated successfully.
